✨[feat] 댓글 등록, 삭제 시 댓글 개수 변화

+++ server/controller/commentController.js
@@ -33,9 +33,11 @@ exports.addComment = async (req, res, next) => {
   });
 
   const message = { message: '댓글 등록이 완료되었습니다!' };
-  const post = Post.findOne({ postId });
+  const post = await Post.findOne({ postId }).then(po => po);
   // 게시물이 있는지 확인
   if (post) {
+    const length = post.comments.length;
+
     const newComment = new Comment({
       postId,
       commentContent,
@@ -44,7 +46,11 @@ exports.addComment = async (req, res, next) => {
     });
     await newComment.save();
 
-    await Post.findOneAndUpdate({ postId }, { $push: { comments: newComment } })
+    await Post.findOneAndUpdate(
+      { postId },
+      // 댓글 등록 시 댓글 개수 추가
+      { $push: { comments: newComment }, commentCount: length + 1 },
+    )
       .then(() => {
         res.status(200).json(message);
       })
@@ -52,32 +58,43 @@ exports.addComment = async (req, res, next) => {
         res.status(500).send(err);
       });
   } else {
-    res.status(500).send({ message: '게시물이 존재하지 않습니다!' });
+    res.status(403).send({ message: '게시물이 존재하지 않습니다!' });
   }
 };
 
 // 댓글 삭제
 exports.deleteComment = async (req, res) => {
   const { commentId } = req.params;
+  const post = await Post.findOne({ postId }).then(po => po);
 
-  // 복호화한 토큰으로 유저 확인
-  let userCheck = await User.findOne({
-    userName: req.tokenInfo,
-  });
-
-  const message = { message: `${commentId}번 댓글이 삭제되었습니다!` };
-
-  await Comment.findOneAndDelete(
-    { commentId },
-    { userName: userCheck.userName },
-  )
-    .then(comment => {
-      if (!comment) {
-        return res.status(500).send({ message: '이미 삭제된 댓글입니다' });
-      }
-      return res.status(200).send(message);
-    })
-    .catch(err => {
-      res.status(500).send(err);
+  // 게시물이 있는지 확인
+  if (post) {
+    // 복호화한 토큰으로 유저 확인
+    let userCheck = await User.findOne({
+      userName: req.tokenInfo,
     });
+
+    const message = { message: `${commentId}번 댓글이 삭제되었습니다!` };
+
+    await Comment.findOneAndDelete(
+      { commentId },
+      { userName: userCheck.userName },
+    )
+      .then(comment => {
+        if (!comment) {
+          return res.status(500).send({ message: '이미 삭제된 댓글입니다' });
+        }
+        Post.findOneAndUpdate(
+          { postId },
+          // 댓글 삭제 시 댓글 개수 감소
+          { commentCount: length - 1 },
+        );
+        return res.status(200).send(message);
+      })
+      .catch(err => {
+        res.status(500).send(err);
+      });
+  } else {
+    res.status(403).send({ message: '게시물이 존재하지 않습니다!' });
+  }
 };
